Add files via upload
I learned that you can use the srand function plus putting "NULL" in the time function in the srand function to initialize a random number each time a 'game' is started. Also to get a range of numbers you can use the modulus and add 1 to not get zero as a number.